Revert "proper subroutes for analyze"
This reverts commit 5fec6b107781961e4772804aabd0cc7fbc2c278c.
