--- web/src/components/RouterWrapper.vue
@@ -1,4 +1,0 @@
-<template lang="pug">
-//- helper component for just showing the route for proper nesting
-router-view
-</template>
